requirements.txt
grpcio
protobuf
grpcio-tools
mypy-protobuf
numpy
torch
pandas
matplotlib
seaborn
scikit-learn